requirements.txt
openai-whisper
sounddevice
PyAudio